requirements.txt
agora-token-builder==1.0.0
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==3.2.4
django-heroku==0.3.1
django-jazzmin==2.5.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
psycopg2==2.9.3
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
whitenoise==6.2.0
